{"schema":"decodex/commit/1","summary":"Modularize autonomy state and app-server boundaries","authority":"manual"}
