+++ apps/decodex/src/orchestrator/status_autonomy/evidence/replay/completeness.rs
@@ -1,0 +1,6 @@
+pub(crate) fn operator_autonomy_evidence_completeness_rank(value: &str) -> u8 {
+	match value {
+		"complete" => 1,
+		_ => 0,
+	}
+}
